Reorganize Menu directory
