+++ SpacePhysics/SpacePhysics/Menu/PauseMenu.cs
@@ -1,11 +1,10 @@
 using System;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
-using Microsoft.Xna.Framework.Input;
-using SpacePhysics.Scenes.Start;
 using SpacePhysics.HUD;
 using static SpacePhysics.GameState;
 using static SpacePhysics.Menu.MenuContainer;
+using SpacePhysics.Menu.MenuItems;
 
 namespace SpacePhysics.Menu;
