+++ SpacePhysics/SpacePhysics/Menu/SubMenus/SoundMenu.cs
@@ -1,17 +1,18 @@
 using System;
 using Microsoft.Xna.Framework;
+using SpacePhysics.Menu.MenuItems;
 using static SpacePhysics.GameState;
 using static SpacePhysics.Menu.MenuContainer;
 
-namespace SpacePhysics.Menu;
+namespace SpacePhysics.Menu.SubMenus;
 
-public class UIMenu : SubMenu
+public class SoundMenu : SubMenu
 {
 
-  public UIMenu() : base(
-    "UI",
-    new Vector2(0f, -50f),
-    State.UI,
+  public SoundMenu() : base(
+    "Sound",
+    new Vector2(0f, -150f),
+    State.Sound,
     State.Settings
   )
   { }
@@ -19,7 +20,7 @@ public class UIMenu : SubMenu
   public override void AddMenuItems()
   {
     menuItems.Add(new ControlItem(
-      "Scale",
+      "Master",
       () => "100%",
       () => activeMenu == 1,
       alignment,
@@ -30,8 +31,8 @@ public class UIMenu : SubMenu
     ));
 
     menuItems.Add(new ControlItem(
-      "Color",
-      () => "Yellow",
+      "Music",
+      () => "100%",
       () => activeMenu == 2,
       alignment,
       () => new Vector2(0f, menuSizeY) + menuOffsetOverride + entireOffsetOverride,
@@ -41,11 +42,22 @@ public class UIMenu : SubMenu
     ));
 
     menuItems.Add(new ControlItem(
-      "Safe zone",
-      () => "0.0",
+      "Sound effects",
+      () => "100%",
       () => activeMenu == 3,
       alignment,
       () => new Vector2(0f, menuSizeY * 2f) + menuOffsetOverride + entireOffsetOverride,
+      controlItemDistance,
+      () => opacity,
+      11
+    ));
+
+    menuItems.Add(new ControlItem(
+      "Menu sound effects",
+      () => "100%",
+      () => activeMenu == 4,
+      alignment,
+      () => new Vector2(0f, menuSizeY * 3f) + menuOffsetOverride + entireOffsetOverride,
       controlItemDistance,
       () => opacity,
       11
