+++ SpacePhysics/SpacePhysics/Menu/SubMenus/UIMenu.cs
@@ -1,17 +1,18 @@
 using System;
 using Microsoft.Xna.Framework;
+using SpacePhysics.Menu.MenuItems;
 using static SpacePhysics.GameState;
 using static SpacePhysics.Menu.MenuContainer;
 
-namespace SpacePhysics.Menu;
+namespace SpacePhysics.Menu.SubMenus;
 
-public class SoundMenu : SubMenu
+public class UIMenu : SubMenu
 {
 
-  public SoundMenu() : base(
-    "Sound",
-    new Vector2(0f, -150f),
-    State.Sound,
+  public UIMenu() : base(
+    "UI",
+    new Vector2(0f, -50f),
+    State.UI,
     State.Settings
   )
   { }
@@ -19,7 +20,7 @@ public class SoundMenu : SubMenu
   public override void AddMenuItems()
   {
     menuItems.Add(new ControlItem(
-      "Master",
+      "Scale",
       () => "100%",
       () => activeMenu == 1,
       alignment,
@@ -30,8 +31,8 @@ public class SoundMenu : SubMenu
     ));
 
     menuItems.Add(new ControlItem(
-      "Music",
-      () => "100%",
+      "Color",
+      () => "Yellow",
       () => activeMenu == 2,
       alignment,
       () => new Vector2(0f, menuSizeY) + menuOffsetOverride + entireOffsetOverride,
@@ -41,22 +42,11 @@ public class SoundMenu : SubMenu
     ));
 
     menuItems.Add(new ControlItem(
-      "Sound effects",
-      () => "100%",
+      "Safe zone",
+      () => "0.0",
       () => activeMenu == 3,
       alignment,
       () => new Vector2(0f, menuSizeY * 2f) + menuOffsetOverride + entireOffsetOverride,
-      controlItemDistance,
-      () => opacity,
-      11
-    ));
-
-    menuItems.Add(new ControlItem(
-      "Menu sound effects",
-      () => "100%",
-      () => activeMenu == 4,
-      alignment,
-      () => new Vector2(0f, menuSizeY * 3f) + menuOffsetOverride + entireOffsetOverride,
       controlItemDistance,
       () => opacity,
       11
